Add basic authentication flow and mocked dashboard
The waitlist signup is removed here and email login/registration
is used instead. The user can login to the dashboard that is
currently mocked. User can also update their profile both in
dashboard when they first login and in the account page
after that.

The users will now have  a "name" field which is a mandatory
field so that we can refer to them by their name, be it a compnay
name or their personal one.

All of the previous waitlist signup related code, including tests,
have been removed here as well. Some miscellaneous UI
facelifts are also included here.