Replace error-prone and unperformant regex matching from route queries
Use `LIKE` with simple % wildcards instead. The trick to avoid the
pattern `/foo/%` of incorrectly matching `/foo/123/bar` is to make sure
that in addition to the `LIKE` match the count of `/` characters in the
strings must be the same.

Also, add an expression index to metrics and dynamic_routes tables which
count the number of `/` occurrences in metrics.path and
dynamic_routes.pattern respectively. Not sure if this makes a lot of
performance difference though, but shouldn't hurt at least since
insertion speed on both tables is much less important than retrieval
speed.

We don't have to worry about doing any data migrations here, since the
only one who is currently has some routes in production's origin_routes
(now renamed to dynamic_routes) is us.