Add various performance improvements
The excluded and dynamic route logic is
changed to such that each metric holds
the knowledge of which dynamic or excluded
route they belong to for faster querying.

Query logic is optimized by utilizing hyper
functions from TimescaleDB and improving
the filtering based on the dynamic and excluded
routes.

The function for fetching metrics for origin is split
up to multiple functions, making it easier to split
querying the metrics to several API endpoints.

Various DB indexes are changed so that all values
that are being used as filters are now combined to
shared indexes for tables.

The origin list now fetches the metrics only for the
last 24 hours to speed up the querying.