Optimize dynamic/excluded routes
This change sets the cascade options on the
`dynamic_route_id` and `excluded_route_id`
fields on the Metric table to `SetNull`, which
handles updating the required metrics whenever
a dynamic/excluded route is deleted.

Another change is done in the queries that
update the metrics for the dynamic/excluded
routes - the metrics are now filtered more
efficiently so that only the matching metrics
are processed in the query.