Fix  `add_metric_endpoint` migration
Need to explicitly set the default value for
new hypertable text columns for some reason
based on local testting.

Also, attempt to include the data migration here
again since that wasn't the original problem.

Also, do not re-create the index as we don'nt need
to filter based on the endpoint after all.